+++ src/Container/Container.php
@@ -50,6 +50,7 @@ use Phalcon\Container\Exceptions\ServiceNotRegistered;
 use Phalcon\Container\Resolver\Lazy\Lazy;
 use Phalcon\Container\Resolver\Resolver;
 use Phalcon\Contracts\Container\Service\Collection;
+use Phalcon\Contracts\Container\Service\Enumerable;
 use Phalcon\Di\InjectionAwareInterface;
 use ReflectionException;
 
@@ -58,7 +59,7 @@ use function class_exists;
 use function in_array;
 use function is_object;
 
-class Container implements Collection
+class Container implements Collection, Enumerable
 {
     /**
      * @var array<string, string>